add custom hook useMyGames
