+++ app/src/Games/index.js
@@ -3,30 +3,33 @@ import * as React from "react";
 import toast, { Toaster } from "react-hot-toast";
 
 import useApi from "../auth/useApi";
+import { useMyGames } from "../hooks";
 
 import styles from "./styles.module.scss";
 
 const Games = () => {
-  const [games, setGames] = React.useState([]);
-  const { loading, apiClient } = useApi();
-
-  const loadGames = React.useCallback(
-    async () => setGames(await apiClient.getGames()),
-    [apiClient],
-  );
-  React.useEffect(() => {
-    !loading && loadGames();
-  }, [loading, loadGames]);
-
-  return loading ? null : (
+  const { myGames } = useMyGames();
+  console.log({ myGames });
+  return myGames ? (
     <>
       <h1>list of all games with remove buttons plus +add a game</h1>
       <section>
-        <GameList {...{ games }} />
+        <GameList games={myGames} />
       </section>
     </>
-  );
+  ) : null;
 };
+
+const GameList = ({ games }) => (
+  <ul className={styles.grid}>
+    {games.map((game) => (
+      <li className={styles.card}>
+        <GameCard {...{ game }} />
+      </li>
+    ))}
+  </ul>
+);
+
 const GameCard = ({ game }) => {
   const onClick = () => {
     // TODO: make this real
@@ -49,14 +52,5 @@ const GameCard = ({ game }) => {
     </>
   );
 };
-const GameList = ({ games }) => (
-  <ul className={styles.grid}>
-    {games.map((game) => (
-      <li className={styles.card}>
-        <GameCard {...{ game }} />
-      </li>
-    ))}
-  </ul>
-);
 
 export default Games;
